requirements.txt
app==0.0.1
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
click==7.1.2
colorama==0.4.4
Django==3.1.7
django-crispy-forms==1.11.1
fastapi==0.63.0
Flask==1.1.2
ghostscript==0.7
gunicorn==20.1.0
h11==0.12.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
magickwand==0.2
MarkupSafe==1.1.1
mccabe==0.6.1
MyApplication==0.1.0
mysqlclient==2.0.3
numpy==1.20.2
opencv-python==4.5.1.48
pdf2image==1.14.0
Pillow==8.2.0
prettytable==2.1.0
pycodestyle==2.7.0
pydantic==1.8.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytesseract==0.3.7
pytz==2021.1
sqlparse==0.4.1
starlette==0.13.6
tabulate==0.8.9
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
Wand==0.6.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1

gensim~=4.1.2
nltk~=3.6.5
requests~=2.25.1
beautifulsoup4~=4.10.0
matplotlib~=3.4.3